+++ packages/ui-react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @betswirl/ui-react
 
+## 0.1.11
+
+### Patch Changes
+
+- Replace gross/raw multiplier/house edge per "BP" type
+
+- Updated dependencies []:
+  - @betswirl/wagmi-provider@0.1.15
+  - @betswirl/sdk-core@0.1.15
+
 ## 0.1.10
 
 ### Patch Changes